relocation 로직 수정 및 status 계산로직 제거

+++ src/main/java/com/winten/greenlight/scheduler/db/repository/redis/customer/CustomerRepository.java
@@ -55,7 +55,7 @@ public class CustomerRepository {
      * @param availableCapacity
      * @return
      */
-    public List<Customer> getTopNCustomersBy(Long actionGroupId, Long availableCapacity) {
+    public List<Customer> getTopNCustomersBy(Long actionGroupId, int availableCapacity) {
         //Set<ZSetOperations.TypedTuple<String>> tuples = zSetOperations.rangeWithScores(WaitingPhase.WAITING.queueName(), 0, eventBackPressure - 1);
         String redisKey = redisKeyBuilder.actionGroupQueue(actionGroupId, WaitStatus.WAITING);
         Set<ZSetOperations.TypedTuple<String>> tuples = zSetOperations.rangeWithScores(redisKey, 0, availableCapacity -1);
